Fixed problems with ScoreboardLib (ColorMatching & Flickering) (Now 122 chars for 1.14+ and 48 chars for 1.13-)
