+++ pom.xml
@@ -6,8 +6,7 @@
 
     <groupId>me.tigerhix.lib</groupId>
     <artifactId>scoreboard</artifactId>
-    <version>1.0.4</version>
-
+    <version>1.0.5</version>
     <packaging>jar</packaging>
 
     <properties>
@@ -53,8 +52,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.0</version>
                 <configuration>
-                    <source>1.7</source>
-                    <target>1.7</target>
+                    <source>8</source>
+                    <target>8</target>
                     <compilerArguments>
                         <O>-Xlint:all</O>
                         <O>-Xlint:-path</O>
