+++ src/main/java/me/tigerhix/lib/scoreboard/ScoreboardLib.java
@@ -1,7 +1,9 @@
 package me.tigerhix.lib.scoreboard;
 
-import me.tigerhix.lib.scoreboard.type.Scoreboard;
 import me.tigerhix.lib.scoreboard.type.SimpleScoreboard;
+import me.tigerhix.lib.scoreboard.type.Scoreboard;
+import me.tigerhix.lib.scoreboard.type.LegacySimpleScoreboard;
+import me.tigerhix.lib.scoreboard.util.ServerVersion;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.Plugin;
 import org.bukkit.plugin.java.JavaPlugin;
@@ -20,7 +22,10 @@ public final class ScoreboardLib extends JavaPlugin {
     }
 
     public static Scoreboard createScoreboard(Player holder) {
-        return new SimpleScoreboard(holder);
+        if (ServerVersion.Version.isCurrentEqualOrHigher(ServerVersion.Version.v1_14_R1)){
+            return new SimpleScoreboard(holder);
+        }
+        return new LegacySimpleScoreboard(holder);
     }
 
     @Override
