+++ src/main/java/me/tigerhix/lib/scoreboard/type/SimpleScoreboard.java
@@ -1,43 +1,33 @@
 package me.tigerhix.lib.scoreboard.type;
 
-import com.google.common.collect.HashBasedTable;
-import com.google.common.collect.Table;
+import java.util.ArrayList;
+import java.util.List;
+
 import me.tigerhix.lib.scoreboard.ScoreboardLib;
 import me.tigerhix.lib.scoreboard.common.Strings;
-import me.tigerhix.lib.scoreboard.util.ServerVersion;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
-import org.bukkit.Location;
-import org.bukkit.OfflinePlayer;
 import org.bukkit.entity.Player;
-import org.bukkit.scheduler.BukkitRunnable;
+import org.bukkit.scheduler.BukkitTask;
 import org.bukkit.scoreboard.DisplaySlot;
 import org.bukkit.scoreboard.Objective;
 import org.bukkit.scoreboard.Team;
 
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
-import java.util.concurrent.ConcurrentHashMap;
-
+/**
+ * @author Tigerpanzer_02
+ * <p>
+ * Created at 28.12.2020
+ */
 public class SimpleScoreboard implements Scoreboard {
 
     private static final String TEAM_PREFIX = "Board_";
-    private static int TEAM_COUNTER = 0;
-
     private final org.bukkit.scoreboard.Scoreboard scoreboard;
     private final Objective objective;
-
     protected Player holder;
     protected long updateInterval = 10L;
-
     private boolean activated;
     private ScoreboardHandler handler;
-    private Map<FakePlayer, Integer> entryCache = new ConcurrentHashMap<>();
-    private Table<String, Integer, FakePlayer> playerCache = HashBasedTable.create();
-    private Table<Team, String, String> teamCache = HashBasedTable.create();
-    private BukkitRunnable updateTask;
+    private BukkitTask updateTask;
 
     public SimpleScoreboard(Player holder) {
         this.holder = holder;
@@ -45,28 +35,32 @@ public class SimpleScoreboard implements Scoreboard {
         scoreboard = Bukkit.getScoreboardManager().getNewScoreboard();
         scoreboard.registerNewObjective("board", "dummy").setDisplaySlot(DisplaySlot.SIDEBAR);
         objective = scoreboard.getObjective(DisplaySlot.SIDEBAR);
+        for (int i = 1; i <= 15; i++) {
+            Team team = scoreboard.registerNewTeam(TEAM_PREFIX + i);
+            team.addEntry(genEntry(i));
+        }
     }
 
     @Override
     public void activate() {
-        if (activated) return;
-        if (handler == null) throw new IllegalArgumentException("Scoreboard handler not set");
+        if (activated) {
+            return;
+        }
+        if (handler == null) {
+            throw new IllegalArgumentException("Scoreboard handler not set");
+        }
         activated = true;
         // Set to the custom scoreboard
         holder.setScoreboard(scoreboard);
         // And start updating on a desired interval
-        updateTask = new BukkitRunnable() {
-            @Override
-            public void run() {
-                update();
-            }
-        };
-        updateTask.runTaskTimer(ScoreboardLib.getPluginInstance(), 0, updateInterval);
+        updateTask = Bukkit.getServer().getScheduler().runTaskTimer(ScoreboardLib.getPluginInstance(), this::update, 0, updateInterval);
     }
 
     @Override
     public void deactivate() {
-        if (!activated) return;
+        if (!activated) {
+            return;
+        }
         activated = false;
         // Set to the main scoreboard
         if (holder.isOnline()) {
@@ -75,7 +69,7 @@ public class SimpleScoreboard implements Scoreboard {
             }
         }
         // Unregister teams that are created for this scoreboard
-        for (Team team : teamCache.rowKeySet()) {
+        for (Team team : scoreboard.getTeams()) {
             team.unregister();
         }
         // Stop updating
@@ -104,8 +98,10 @@ public class SimpleScoreboard implements Scoreboard {
     }
 
     @Override
-    public SimpleScoreboard setUpdateInterval(long updateInterval) {
-        if (activated) throw new IllegalStateException("Scoreboard is already activated");
+    public Scoreboard setUpdateInterval(long updateInterval) {
+        if (activated) {
+            throw new IllegalStateException("Scoreboard is already activated");
+        }
         this.updateInterval = updateInterval;
         return this;
     }
@@ -115,7 +111,6 @@ public class SimpleScoreboard implements Scoreboard {
         return holder;
     }
 
-    @SuppressWarnings("deprecation")
     private void update() {
         if (!holder.isOnline()) {
             deactivate();
@@ -123,253 +118,68 @@ public class SimpleScoreboard implements Scoreboard {
         }
         // Title
         String handlerTitle = handler.getTitle(holder);
-        String finalTitle = Strings.format(handlerTitle != null ? handlerTitle : ChatColor.BOLD.toString());
-        if (!objective.getDisplayName().equals(finalTitle)) objective.setDisplayName(Strings.format(finalTitle));
+        String finalTitle = Strings
+                .format(handlerTitle != null ? handlerTitle : ChatColor.BOLD.toString());
+        if (!objective.getDisplayName().equals(finalTitle)) {
+            objective.setDisplayName(Strings.format(finalTitle));
+        }
         // Entries
         List<Entry> passed = handler.getEntries(holder);
-        Map<String, Integer> appeared = new HashMap<>();
-        Map<FakePlayer, Integer> current = new HashMap<>();
-        if (passed == null) return;
+        List<Integer> current = new ArrayList<>();
+        if (passed == null) {
+            return;
+        }
         for (Entry entry : passed) {
             // Handle the entry
             String key = entry.getName();
-            Integer score = entry.getPosition();
-            String appearance;
-            if (ServerVersion.Version.isCurrentEqualOrHigher(ServerVersion.Version.v1_14_R1) && key.length() > 48) {
-                if (key.length() > 144) key = key.substring(0, 143);
-                if (key.length() > 64) {
-                    appearance = key.substring(64);
-                } else {
-                    appearance = key;
-                }
-            } else {
-                if (key.length() > 48) key = key.substring(0, 47);
-                if (key.length() > 16) {
-                    appearance = key.substring(16);
-                } else {
-                    appearance = key;
-                }
+            int score = entry.getPosition();
+
+            Team team = scoreboard.getTeam(TEAM_PREFIX + score);
+            String temp = genEntry(score);
+            if (!scoreboard.getEntries().contains(temp)) {
+                objective.getScore(temp).setScore(score);
             }
-            if (!appeared.containsKey(appearance)) appeared.put(appearance, -1);
-            appeared.put(appearance, appeared.get(appearance) + 1);
-            // Get fake player
-            FakePlayer faker = getFakePlayer(key, appeared.get(appearance));
-            // Set score
-            objective.getScore(faker).setScore(score);
-            // Update references
-            entryCache.put(faker, score);
-            current.put(faker, score);
+
+            // Add prefix & suffix
+            key = Strings.format(key);
+            String prefix = getFirstSplit(key);
+            String suffix = getFirstSplit(ChatColor.getLastColors(prefix) + (prefix.endsWith("§") ? "§" : "") + getSecondSplit(key));
+            team.setPrefix(prefix);
+            team.setSuffix(suffix);
+            // Add to current
+            current.add(score);
         }
-        appeared.clear();
         // Remove duplicated or non-existent entries
-        for (FakePlayer fakePlayer : entryCache.keySet()) {
-            if (!current.containsKey(fakePlayer)) {
-                entryCache.remove(fakePlayer);
-                scoreboard.resetScores(fakePlayer.getName());
-            }
-        }
-    }
-
-    @SuppressWarnings("deprecation")
-    private FakePlayer getFakePlayer(String text, int offset) {
-        Team team = null;
-        String name;
-        // If the text has a length less than 16, teams need not to be be created
-        if (text.length() <= 16) {
-            name = text + Strings.repeat(" ", offset);
-        } else {
-            String prefix;
-            String suffix = "";
-            offset++;
-            // Otherwise, iterate through the string and cut off prefix and suffix
-            if (ServerVersion.Version.isCurrentEqualOrHigher(ServerVersion.Version.v1_14_R1) && text.length() > 48) {
-                if (text.length() > 63) {
-                    prefix = text.substring(0, 64 - offset);
-                    int charRemover = 0;
-                    if (prefix.endsWith("§")) {
-                        charRemover++;
-                    }
-                    name = getLastColor(prefix) + (prefix.endsWith("§") ? "§" : "") + text.substring(64 - offset);
-                    if (name.length() > 16) name = name.substring(0, 16);
-                    // -2 because of getLastColors need 2 chars
-                    if (text.length() > 80)
-                        suffix = getLastColor(name) + (name.endsWith("§") ? "§" : "") + text.substring(80 - 2 - charRemover - offset);
-                    if (suffix.length() > 64) suffix = suffix.substring(0, 64);
-                } else {
-                    prefix = text.substring(0, text.length() - offset);
-                    name = getLastColor(prefix) + (prefix.endsWith("§") ? "§" : "") + text.substring(text.length() - offset);
-                    if (name.length() > 16) name = name.substring(0, 16);
-                }
-            } else {
-                prefix = text.substring(0, 16 - offset);
-                name = getLastColor(prefix) + (prefix.endsWith("§") ? "§" : "") + text.substring(16 - offset);
-                if (name.length() > 16) name = name.substring(0, 16);
-                // -2 because of getLastColors need 2 chars
-                int charRemover = 0;
-                if (prefix.endsWith("§")) {
-                    charRemover++;
-                }
-                if (text.length() > 32)
-                    suffix = getLastColor(name) + (name.endsWith("§") ? "§" : "") + text.substring(32 - 2 - charRemover - offset);
-                if (suffix.length() > 16) suffix = suffix.substring(0, 16);
-            }
-            // If teams already exist, use them
-            for (Team other : teamCache.rowKeySet()) {
-                if (other.getPrefix().equals(prefix) && other.getSuffix().equals(suffix)) {
-                    team = other;
+        for (int i = 1; i <= 15; i++) {
+            if (!current.contains(i)) {
+                String entry = genEntry(i);
+                if (scoreboard.getEntries().contains(entry)) {
+                    scoreboard.resetScores(entry);
                 }
             }
-            // Otherwise create them
-            if (team == null) {
-                team = scoreboard.registerNewTeam(TEAM_PREFIX + TEAM_COUNTER++);
-                team.setPrefix(prefix);
-                team.setSuffix(suffix);
-                teamCache.put(team, prefix, suffix);
-            }
         }
-        FakePlayer faker;
-        if (!playerCache.contains(name, offset)) {
-            faker = new FakePlayer(name, team, offset);
-            playerCache.put(name, offset, faker);
-            if (faker.getTeam() != null) {
-                faker.getTeam().addPlayer(faker);
-            }
-        } else {
-            faker = playerCache.get(name, offset);
-            if (team != null && faker.getTeam() != null) {
-                faker.getTeam().removePlayer(faker);
-            }
-            faker.setTeam(team);
-            if (faker.getTeam() != null) {
-                faker.getTeam().addPlayer(faker);
-            }
-        }
-        return faker;
-    }
-
-    private String getLastColor(String message) {
-        String color = ChatColor.getLastColors(message);
-        if (color.equalsIgnoreCase("")) {
-            return "§f";
-        }
-        return color;
     }
 
     public Objective getObjective() {
         return objective;
     }
 
+    private String genEntry(int slot) {
+        return ChatColor.values()[slot].toString();
+    }
+
     public org.bukkit.scoreboard.Scoreboard getScoreboard() {
         return scoreboard;
     }
 
-    private static class FakePlayer implements OfflinePlayer {
-
-        private final String name;
-
-        private Team team;
-        private int offset;
-
-        FakePlayer(String name, Team team, int offset) {
-            this.name = name;
-            this.team = team;
-            this.offset = offset;
-        }
-
-        public Team getTeam() {
-            return team;
-        }
-
-        public void setTeam(Team team) {
-            this.team = team;
-        }
-
-        public int getOffset() {
-            return offset;
-        }
-
-        public String getFullName() {
-            if (team == null) return name;
-            if (team.getSuffix() == null) return team.getPrefix() + name;
-            return team.getPrefix() + name + team.getSuffix();
-        }
-
-        @Override
-        public boolean isOnline() {
-            return true;
-        }
-
-        @Override
-        public String getName() {
-            return name;
-        }
-
-        @Override
-        public UUID getUniqueId() {
-            return UUID.randomUUID();
-        }
-
-        @Override
-        public boolean isBanned() {
-            return false;
-        }
-
-        @Override
-        public boolean isWhitelisted() {
-            return false;
-        }
-
-        @Override
-        public void setWhitelisted(boolean whitelisted) {
-        }
-
-        @Override
-        public Player getPlayer() {
-            return null;
-        }
-
-        @Override
-        public long getFirstPlayed() {
-            return 0;
-        }
-
-        @Override
-        public long getLastPlayed() {
-            return 0;
-        }
-
-        @Override
-        public boolean hasPlayedBefore() {
-            return false;
-        }
-
-        @Override
-        public Location getBedSpawnLocation() {
-            return null;
-        }
-
-        @Override
-        public Map<String, Object> serialize() {
-            return null;
-        }
-
-        @Override
-        public boolean isOp() {
-            return false;
-        }
-
-        @Override
-        public void setOp(boolean op) {
-        }
-
-        @Override
-        public String toString() {
-            return "FakePlayer{" +
-                    "name='" + name + '\'' +
-                    ", team=" + team
-                    + '}';
-        }
-
+    private String getFirstSplit(String s) {
+        return s.length() > 64 ? s.substring(0, 64) : s;
     }
 
+    private String getSecondSplit(String s) {
+        if (s.length() > 128) {
+            s = s.substring(0, 128);
+        }
+        return s.length() > 64 ? s.substring(64) : "";
+    }
 }
